requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1
pydantic>=2.5.0
python-multipart>=0.0.6
newspaper3k>=0.2.8
beautifulsoup4>=4.12.2
transformers>=4.35.2
torch>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
bleach>=6.1.0
asyncpg>=0.29.0
aiohttp>=3.9.1